NewsPaper/requirements.txt
Django
django-allauth
django-filter
celery
redis
django-celery-beat
